+++ attention/sparse_flash_attention/op_host/sparse_flash_attention_tiling.cpp
@@ -696,7 +696,7 @@ ge::graphStatus SFATilingCheck::CheckSingleParaSparseMode() const
 ge::graphStatus SFATilingCheck::CheckSingleParaSparseBlockSize() const
 {
     OP_CHECK_IF((*opParamInfo_.sparseBlockSize <= 0 || *opParamInfo_.sparseBlockSize > 128 ||
-        (*opParamInfo_.sparseBlockSize & (*opParamInfo_.sparseBlockSize - 1)) != 0),
+        (static_cast<uint64_t>(*opParamInfo_.sparseBlockSize) & static_cast<uint64_t>(*opParamInfo_.sparseBlockSize - 1L)) != 0UL),
         OP_LOGE(opName_, "sparseBlockSize should be be in range [1, 128] and be a power of 2, but got: %ld.",
             *opParamInfo_.sparseBlockSize),
         return ge::GRAPH_FAILED);
